refactor: optimize the directory structure of dit.
