+++ xllm/models/dit/pipelines/pipeline_flux.h
@@ -14,33 +14,33 @@ limitations under the License.
 ==============================================================================*/
 
 #pragma once
-#include "pipeline_flux_base.h"
-#include "transformer_flux.h"
-// pipeline_flux_control compatible with huggingface weights
+#include "models/dit/pipelines/pipeline_flux_base.h"
+#include "models/dit/transformers/transformer_flux.h"
+// pipeline_flux compatible with huggingface weights
 // ref to:
-// https://github.com/huggingface/diffusers/blob/main/src/diffusers/pipelines/flux/pipeline_flux_control.py
+// https://github.com/huggingface/diffusers/blob/main/src/diffusers/pipelines/flux/pipeline_flux.py
 
 namespace xllm {
 
-class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
+class FluxPipelineImpl : public FluxPipelineBaseImpl {
  public:
-  FluxControlPipelineImpl(const DiTModelContext& context) {
-    auto model_args = context.get_model_args("vae");
+  FluxPipelineImpl(const DiTModelContext& context) {
+    const auto& model_args = context.get_model_args("vae");
     options_ = context.get_tensor_options();
     vae_scale_factor_ = 1 << (model_args.block_out_channels().size() - 1);
+
     vae_shift_factor_ = model_args.shift_factor();
     vae_scaling_factor_ = model_args.scale_factor();
-    latent_channels_ = model_args.latent_channels();
     tokenizer_max_length_ =
         context.get_model_args("text_encoder").max_position_embeddings();
-    LOG(INFO) << "Initializing FluxControl pipeline...";
-    image_processor_ = VAEImageProcessor(context.get_model_context("vae"),
-                                         true,
-                                         true,
-                                         false,
-                                         false,
-                                         false,
-                                         latent_channels_);
+    LOG(INFO) << "Initializing Flux pipeline...";
+    vae_image_processor_ = VAEImageProcessor(context.get_model_context("vae"),
+                                             true,
+                                             true,
+                                             false,
+                                             false,
+                                             false,
+                                             model_args.latent_channels());
     vae_ = VAE(context.get_model_context("vae"));
     pos_embed_ = register_module(
         "pos_embed",
@@ -52,7 +52,7 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
     scheduler_ =
         FlowMatchEulerDiscreteScheduler(context.get_model_context("scheduler"));
     register_module("vae", vae_);
-    register_module("vae_image_processor", image_processor_);
+    register_module("vae_image_processor", vae_image_processor_);
     register_module("transformer", transformer_);
     register_module("t5", t5_);
     register_module("scheduler", scheduler_);
@@ -61,40 +61,57 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
 
   DiTForwardOutput forward(const DiTForwardInput& input) {
     const auto& generation_params = input.generation_params;
-    int64_t height = generation_params.height;
-    int64_t width = generation_params.width;
+
     auto seed = generation_params.seed > 0 ? generation_params.seed : 42;
     auto prompts = std::make_optional(input.prompts);
     auto prompts_2 = input.prompts_2.empty()
                          ? std::nullopt
                          : std::make_optional(input.prompts_2);
-
-    auto control_image = input.control_image;
+    auto negative_prompts = input.negative_prompts.empty()
+                                ? std::nullopt
+                                : std::make_optional(input.negative_prompts);
+    auto negative_prompts_2 =
+        input.negative_prompts_2.empty()
+            ? std::nullopt
+            : std::make_optional(input.negative_prompts_2);
 
     auto latents = input.latents.defined() ? std::make_optional(input.latents)
                                            : std::nullopt;
     auto prompt_embeds = input.prompt_embeds.defined()
                              ? std::make_optional(input.prompt_embeds)
                              : std::nullopt;
+    auto negative_prompt_embeds =
+        input.negative_prompt_embeds.defined()
+            ? std::make_optional(input.negative_prompt_embeds)
+            : std::nullopt;
     auto pooled_prompt_embeds =
         input.pooled_prompt_embeds.defined()
             ? std::make_optional(input.pooled_prompt_embeds)
             : std::nullopt;
+    auto negative_pooled_prompt_embeds =
+        input.negative_pooled_prompt_embeds.defined()
+            ? std::make_optional(input.negative_pooled_prompt_embeds)
+            : std::nullopt;
 
-    auto output = forward_impl(prompts,
-                               prompts_2,
-                               control_image,
-                               height,
-                               width,
-                               generation_params.strength,
-                               generation_params.num_inference_steps,
-                               generation_params.guidance_scale,
-                               generation_params.num_images_per_prompt,
-                               seed,
-                               latents,
-                               prompt_embeds,
-                               pooled_prompt_embeds,
-                               generation_params.max_sequence_length);
+    auto output = forward_impl(
+        prompts,                                  // prompt
+        prompts_2,                                // prompt_2
+        negative_prompts,                         // negative_prompt
+        negative_prompts_2,                       // negative_prompt_2
+        generation_params.true_cfg_scale,         // cfg scale
+        generation_params.height,                 // height
+        generation_params.width,                  // width
+        generation_params.num_inference_steps,    // num_inference_steps
+        generation_params.guidance_scale,         // guidance_scale
+        generation_params.num_images_per_prompt,  // num_images_per_prompt
+        seed,                                     // seed
+        latents,                                  // latents
+        prompt_embeds,                            // prompt_embeds
+        negative_prompt_embeds,                   // negative_prompt_embeds
+        pooled_prompt_embeds,                     // pooled_prompt_embeds
+        negative_pooled_prompt_embeds,         // negative_pooled_prompt_embeds
+        generation_params.max_sequence_length  // max_sequence_length
+    );
 
     DiTForwardOutput out;
     out.tensors = torch::chunk(output, input.batch_size);
@@ -102,8 +119,7 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
   }
 
   void load_model(std::unique_ptr<DiTModelLoader> loader) {
-    LOG(INFO) << "FluxControlPipeline loading model from"
-              << loader->model_root_path();
+    LOG(INFO) << "FluxPipeline loading model from" << loader->model_root_path();
     std::string model_path = loader->model_root_path();
     auto transformer_loader = loader->take_component_loader("transformer");
     auto vae_loader = loader->take_component_loader("vae");
@@ -112,8 +128,7 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
     auto tokenizer_loader = loader->take_component_loader("tokenizer");
     auto tokenizer_2_loader = loader->take_component_loader("tokenizer_2");
     LOG(INFO)
-        << "FluxControl model components loaded, start to load weights to "
-           "sub models";
+        << "Flux model components loaded, start to load weights to sub models";
     transformer_->load_model(std::move(transformer_loader));
     transformer_->to(options_.device());
     vae_->load_model(std::move(vae_loader));
@@ -127,25 +142,6 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
   }
 
  private:
-  torch::Tensor encode_vae_image(const torch::Tensor& image, int64_t seed) {
-    torch::Tensor latents = vae_->encode(image, seed);
-    latents = (latents - vae_shift_factor_) * vae_scaling_factor_;
-    return latents;
-  }
-
-  std::pair<torch::Tensor, int64_t> get_timesteps(int64_t num_inference_steps,
-                                                  float strength) {
-    int64_t init_timestep =
-        std::min(static_cast<int64_t>(num_inference_steps * strength),
-                 num_inference_steps);
-
-    int64_t t_start = std::max(num_inference_steps - init_timestep, int64_t(0));
-    int64_t start_idx = t_start * scheduler_->order();
-    auto timesteps = scheduler_->timesteps().slice(0, start_idx).to(options_);
-    scheduler_->set_begin_index(start_idx);
-    return {timesteps, num_inference_steps - t_start};
-  }
-
   std::pair<torch::Tensor, torch::Tensor> prepare_latents(
       int64_t batch_size,
       int64_t num_channels_latents,
@@ -174,49 +170,25 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
     return {packed_latents, latent_image_ids};
   }
 
-  torch::Tensor prepare_image(torch::Tensor image,
-                              int64_t width,
-                              int64_t height,
-                              int64_t batch_size,
-                              int64_t num_images_per_prompt) {
-    int image_batch_size = image.size(0);
-    int repeat_times;
-    image = image_processor_->preprocess(
-        image, height, width, "default", std::nullopt);
-    if (image_batch_size == 1) {
-      repeat_times = batch_size;
-    } else {
-      repeat_times = num_images_per_prompt;
-    }
-    const auto B = image.size(0);
-    const auto C = image.size(1);
-    const auto H = image.size(2);
-    const auto W = image.size(3);
-    image = image.unsqueeze(1)
-                .repeat({1, repeat_times, 1, 1, 1})
-                .reshape({B * repeat_times, C, H, W})
-                .to(options_);
-    return image;
-  }
-
   torch::Tensor forward_impl(
       std::optional<std::vector<std::string>> prompt = std::nullopt,
       std::optional<std::vector<std::string>> prompt_2 = std::nullopt,
-      torch::Tensor control_image = torch::Tensor(),
+      std::optional<std::vector<std::string>> negative_prompt = std::nullopt,
+      std::optional<std::vector<std::string>> negative_prompt_2 = std::nullopt,
+      float true_cfg_scale = 1.0f,
       int64_t height = 512,
       int64_t width = 512,
-      float strength = 1.0f,
-      int64_t num_inference_steps = 50,
-      float guidance_scale = 30.0f,
+      int64_t num_inference_steps = 28,
+      float guidance_scale = 3.5f,
       int64_t num_images_per_prompt = 1,
-      int64_t seed = 42,
+      std::optional<int64_t> seed = std::nullopt,
       std::optional<torch::Tensor> latents = std::nullopt,
       std::optional<torch::Tensor> prompt_embeds = std::nullopt,
       std::optional<torch::Tensor> pooled_prompt_embeds = std::nullopt,
+      std::optional<torch::Tensor> negative_prompt_embeds = std::nullopt,
+      std::optional<torch::Tensor> negative_pooled_prompt_embeds = std::nullopt,
       int64_t max_sequence_length = 512) {
     torch::NoGradGuard no_grad;
-    int64_t actual_height = height;
-    int64_t actual_width = width;
     int64_t batch_size;
     if (prompt.has_value()) {
       batch_size = prompt.value().size();
@@ -224,6 +196,10 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
       batch_size = prompt_embeds.value().size(0);
     }
     int64_t total_batch_size = batch_size * num_images_per_prompt;
+    bool has_neg_prompt = negative_prompt.has_value() ||
+                          (negative_prompt_embeds.has_value() &&
+                           negative_pooled_prompt_embeds.has_value());
+    bool do_true_cfg = (true_cfg_scale > 1.0f) && has_neg_prompt;
     // encode prompt
     auto [encoded_prompt_embeds, encoded_pooled_embeds, text_ids] =
         encode_prompt(prompt,
@@ -232,34 +208,29 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
                       pooled_prompt_embeds,
                       num_images_per_prompt,
                       max_sequence_length);
-
-    // prepare latent
-    int64_t num_channels_latents = transformer_->in_channels() / 8;
-    // control image to latents
-    control_image = prepare_image(control_image,
-                                  width,
-                                  height,
-                                  batch_size * num_images_per_prompt,
-                                  num_images_per_prompt);
-    if (control_image.dim() == 4) {
-      auto enc = vae_->encode(control_image, seed);
-      control_image = (enc - vae_shift_factor_) * vae_scaling_factor_;
-      control_image = control_image.to(options_);
-      auto shape = control_image.sizes();
-      auto height_control_image = shape[2];
-      auto width_control_image = shape[3];
-      control_image = pack_latents(control_image,
-                                   total_batch_size,
-                                   num_channels_latents,
-                                   height_control_image,
-                                   width_control_image);
+    // encode negative prompt
+    torch::Tensor negative_encoded_embeds, negative_pooled_embeds;
+    torch::Tensor negative_text_ids;
+    if (do_true_cfg) {
+      auto [neg_emb, neg_pooled, neg_ids] =
+          encode_prompt(negative_prompt,
+                        negative_prompt_2,
+                        negative_prompt_embeds,
+                        negative_pooled_prompt_embeds,
+                        num_images_per_prompt,
+                        max_sequence_length);
+      negative_encoded_embeds = neg_emb;
+      negative_pooled_embeds = neg_pooled;
+      negative_text_ids = neg_ids;
     }
+    // prepare latent
+    int64_t num_channels_latents = transformer_->in_channels() / 4;
     auto [prepared_latents, latent_image_ids] =
         prepare_latents(total_batch_size,
                         num_channels_latents,
-                        actual_height,
-                        actual_width,
-                        seed,
+                        height,
+                        width,
+                        seed.has_value() ? seed.value() : 42,
                         latents);
     // prepare timestep
     std::vector<float> new_sigmas;
@@ -284,8 +255,10 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
     // prepare guidance
     torch::Tensor guidance;
     if (transformer_->guidance_embeds()) {
-      torch::TensorOptions options = options_.dtype(torch::kFloat32);
-      guidance = torch::full(at::IntArrayRef({1}), guidance_scale, options);
+      torch::TensorOptions options =
+          torch::dtype(torch::kFloat32).device(options_.device());
+
+      guidance = torch::full(torch::IntArrayRef({1}), guidance_scale, options);
       guidance = guidance.expand({prepared_latents.size(0)});
     }
     scheduler_->set_begin_index(0);
@@ -299,21 +272,33 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
                                   width / (vae_scale_factor_ * 2));
     torch::Tensor image_rotary_emb =
         torch::stack({rot_emb1, rot_emb2}, 0).to(options_.dtype());
+
     for (int64_t i = 0; i < timesteps.numel(); ++i) {
       torch::Tensor t = timesteps[i].unsqueeze(0);
       timestep.fill_(t.item<float>())
           .to(prepared_latents.dtype())
           .div_(1000.0f);
       int64_t step_id = i + 1;
-      auto controlled_latents =
-          torch::cat({prepared_latents, control_image}, 2);
-      torch::Tensor noise_pred = transformer_->forward(controlled_latents,
+      torch::Tensor noise_pred = transformer_->forward(prepared_latents,
                                                        encoded_prompt_embeds,
                                                        encoded_pooled_embeds,
                                                        timestep,
                                                        image_rotary_emb,
                                                        guidance,
                                                        step_id);
+      if (do_true_cfg) {
+        torch::Tensor negative_noise_pred =
+            transformer_->forward(prepared_latents,
+                                  negative_encoded_embeds,
+                                  negative_pooled_embeds,
+                                  timestep,
+                                  image_rotary_emb,
+                                  guidance,
+                                  step_id);
+        noise_pred =
+            noise_pred + (noise_pred - negative_noise_pred) * true_cfg_scale;
+        negative_noise_pred.reset();
+      }
       auto prev_latents = scheduler_->step(noise_pred, t, prepared_latents);
       prepared_latents = prev_latents.detach();
       std::vector<torch::Tensor> tensors = {prepared_latents, noise_pred};
@@ -327,28 +312,26 @@ class FluxControlPipelineImpl : public FluxPipelineBaseImpl {
     }
     torch::Tensor image;
     // Unpack latents
-    torch::Tensor unpacked_latents = unpack_latents(
-        prepared_latents, actual_height, actual_width, vae_scale_factor_);
+    torch::Tensor unpacked_latents =
+        unpack_latents(prepared_latents, height, width, vae_scale_factor_);
     unpacked_latents =
         (unpacked_latents / vae_scaling_factor_) + vae_shift_factor_;
     unpacked_latents = unpacked_latents.to(options_.dtype());
     image = vae_->decode(unpacked_latents);
-    image = image_processor_->postprocess(image);
+    image = vae_image_processor_->postprocess(image);
     return image;
   }
 
  private:
   FlowMatchEulerDiscreteScheduler scheduler_{nullptr};
   VAE vae_{nullptr};
-  VAEImageProcessor image_processor_{nullptr};
+  VAEImageProcessor vae_image_processor_{nullptr};
   FluxDiTModel transformer_{nullptr};
   float vae_scaling_factor_;
   float vae_shift_factor_;
-  int64_t vae_latent_channels_;
-  int64_t latent_channels_;
   FluxPosEmbed pos_embed_{nullptr};
 };
-TORCH_MODULE(FluxControlPipeline);
+TORCH_MODULE(FluxPipeline);
 
-REGISTER_DIT_MODEL(flux_control, FluxControlPipeline);
+REGISTER_DIT_MODEL(flux, FluxPipeline);
 }  // namespace xllm
