+++ xllm/models/dit/pipelines/pipeline_longcat_image.h
@@ -20,7 +20,6 @@ limitations under the License.
 #include <memory>
 #include <string>
 
-#include "autoencoder_kl.h"
 #include "core/framework/dit_model_loader.h"
 #include "core/framework/model_context.h"
 #include "core/framework/parallel_state/parallel_args.h"
@@ -28,10 +27,11 @@ limitations under the License.
 #include "core/framework/request/dit_request_state.h"
 #include "core/framework/tokenizer/tokenizer.h"
 #include "core/layers/cuda/flashinfer_workspace.h"
-#include "flowmatch_euler_discrete_scheduler.h"
+#include "models/dit/autoencoders/autoencoder_kl.h"
+#include "models/dit/schedulers/flowmatch_euler_discrete_scheduler.h"
+#include "models/dit/transformers/transformer_longcat_image.h"
 #include "models/model_registry.h"
 #include "models/vlm/qwen2_5_vl.h"
-#include "transformer_longcat_image.h"
 
 namespace xllm {
