+++ xllm/models/models.h
@@ -16,89 +16,89 @@ limitations under the License.
 #pragma once
 
 #if defined(USE_NPU)
-#include "dit/npu/qwen_image_edit/pipeline_qwenimage_edit_plus.h"  // IWYU pragma: keep
-#include "dit/pipeline_flux.h"                // IWYU pragma: keep
-#include "dit/pipeline_flux_control.h"        // IWYU pragma: keep
-#include "dit/pipeline_flux_fill.h"           // IWYU pragma: keep
-#include "llm/deepseek_v4.h"                  // IWYU pragma: keep
-#include "llm/npu/deepseek_mtp.h"             // IWYU pragma: keep
-#include "llm/npu/deepseek_v2.h"              // IWYU pragma: keep
-#include "llm/npu/deepseek_v3.h"              // IWYU pragma: keep
-#include "llm/npu/deepseek_v32.h"             // IWYU pragma: keep
-#include "llm/npu/deepseek_v32_mtp.h"         // IWYU pragma: keep
-#include "llm/npu/glm4.h"                     // IWYU pragma: keep
-#include "llm/npu/glm4_moe.h"                 // IWYU pragma: keep
-#include "llm/npu/glm4_moe_lite.h"            // IWYU pragma: keep
-#include "llm/npu/glm4_moe_mtp.h"             // IWYU pragma: keep
-#include "llm/npu/glm5_moe.h"                 // IWYU pragma: keep
-#include "llm/npu/glm5_moe_mtp.h"             // IWYU pragma: keep
-#include "llm/npu/joyai_llm_flash.h"          // IWYU pragma: keep
-#include "llm/npu/kimi_k2.h"                  // IWYU pragma: keep
-#include "llm/npu/llama.h"                    // IWYU pragma: keep
-#include "llm/npu/llama3.h"                   // IWYU pragma: keep
-#include "llm/npu/minimax_m2.h"               // IWYU pragma: keep
-#include "llm/npu/oxygen.h"                   // IWYU pragma: keep
-#include "llm/npu/qwen2.h"                    // IWYU pragma: keep
-#include "llm/npu/qwen3.h"                    // IWYU pragma: keep
-#include "llm/npu/qwen3_eagle3.h"             // IWYU pragma: keep
-#include "llm/npu/qwen3_moe.h"                // IWYU pragma: keep
-#include "llm/qwen3.h"                        // IWYU pragma: keep
-#include "llm/qwen3_5.h"                      // IWYU pragma: keep
-#include "llm/qwen3_5_mtp.h"                  // IWYU pragma: keep
-#include "llm/qwen3_moe.h"                    // IWYU pragma: keep
-#include "llm/qwen3_next.h"                   // IWYU pragma: keep
-#include "rec/npu/onerec.h"                   // IWYU pragma: keep
-#include "vlm/npu/glm4v.h"                    // IWYU pragma: keep
-#include "vlm/npu/glm4v_moe.h"                // IWYU pragma: keep
-#include "vlm/npu/kimi_k25.h"                 // IWYU pragma: keep
-#include "vlm/npu/minicpmv.h"                 // IWYU pragma: keep
-#include "vlm/npu/oxygen_vlm.h"               // IWYU pragma: keep
-#include "vlm/npu/qwen2_5_vl.h"               // IWYU pragma: keep
-#include "vlm/npu/qwen2_5_vl_mm_embedding.h"  // IWYU pragma: keep
-#include "vlm/npu/qwen2_vl.h"                 // IWYU pragma: keep
-#include "vlm/npu/qwen3_vl.h"                 // IWYU pragma: keep
-#include "vlm/npu/qwen3_vl_mm_embedding.h"    // IWYU pragma: keep
-#include "vlm/npu/qwen3_vl_moe.h"             // IWYU pragma: keep
-#include "vlm/qwen3_5.h"                      // IWYU pragma: keep
+#include "dit/pipelines/pipeline_flux.h"                 // IWYU pragma: keep
+#include "dit/pipelines/pipeline_flux_control.h"         // IWYU pragma: keep
+#include "dit/pipelines/pipeline_flux_fill.h"            // IWYU pragma: keep
+#include "dit/pipelines/pipeline_qwenimage_edit_plus.h"  // IWYU pragma: keep
+#include "llm/deepseek_v4.h"                             // IWYU pragma: keep
+#include "llm/npu/deepseek_mtp.h"                        // IWYU pragma: keep
+#include "llm/npu/deepseek_v2.h"                         // IWYU pragma: keep
+#include "llm/npu/deepseek_v3.h"                         // IWYU pragma: keep
+#include "llm/npu/deepseek_v32.h"                        // IWYU pragma: keep
+#include "llm/npu/deepseek_v32_mtp.h"                    // IWYU pragma: keep
+#include "llm/npu/glm4.h"                                // IWYU pragma: keep
+#include "llm/npu/glm4_moe.h"                            // IWYU pragma: keep
+#include "llm/npu/glm4_moe_lite.h"                       // IWYU pragma: keep
+#include "llm/npu/glm4_moe_mtp.h"                        // IWYU pragma: keep
+#include "llm/npu/glm5_moe.h"                            // IWYU pragma: keep
+#include "llm/npu/glm5_moe_mtp.h"                        // IWYU pragma: keep
+#include "llm/npu/joyai_llm_flash.h"                     // IWYU pragma: keep
+#include "llm/npu/kimi_k2.h"                             // IWYU pragma: keep
+#include "llm/npu/llama.h"                               // IWYU pragma: keep
+#include "llm/npu/llama3.h"                              // IWYU pragma: keep
+#include "llm/npu/minimax_m2.h"                          // IWYU pragma: keep
+#include "llm/npu/oxygen.h"                              // IWYU pragma: keep
+#include "llm/npu/qwen2.h"                               // IWYU pragma: keep
+#include "llm/npu/qwen3.h"                               // IWYU pragma: keep
+#include "llm/npu/qwen3_eagle3.h"                        // IWYU pragma: keep
+#include "llm/npu/qwen3_moe.h"                           // IWYU pragma: keep
+#include "llm/qwen3.h"                                   // IWYU pragma: keep
+#include "llm/qwen3_5.h"                                 // IWYU pragma: keep
+#include "llm/qwen3_5_mtp.h"                             // IWYU pragma: keep
+#include "llm/qwen3_moe.h"                               // IWYU pragma: keep
+#include "llm/qwen3_next.h"                              // IWYU pragma: keep
+#include "rec/npu/onerec.h"                              // IWYU pragma: keep
+#include "vlm/npu/glm4v.h"                               // IWYU pragma: keep
+#include "vlm/npu/glm4v_moe.h"                           // IWYU pragma: keep
+#include "vlm/npu/kimi_k25.h"                            // IWYU pragma: keep
+#include "vlm/npu/minicpmv.h"                            // IWYU pragma: keep
+#include "vlm/npu/oxygen_vlm.h"                          // IWYU pragma: keep
+#include "vlm/npu/qwen2_5_vl.h"                          // IWYU pragma: keep
+#include "vlm/npu/qwen2_5_vl_mm_embedding.h"             // IWYU pragma: keep
+#include "vlm/npu/qwen2_vl.h"                            // IWYU pragma: keep
+#include "vlm/npu/qwen3_vl.h"                            // IWYU pragma: keep
+#include "vlm/npu/qwen3_vl_mm_embedding.h"               // IWYU pragma: keep
+#include "vlm/npu/qwen3_vl_moe.h"                        // IWYU pragma: keep
+#include "vlm/qwen3_5.h"                                 // IWYU pragma: keep
 
 #elif defined(USE_MLU)
-#include "dit/pipeline_flux.h"          // IWYU pragma: keep
-#include "dit/pipeline_flux_control.h"  // IWYU pragma: keep
-#include "dit/pipeline_flux_fill.h"     // IWYU pragma: keep
-#include "llm/deepseek_mtp.h"           // IWYU pragma: keep
-#include "llm/deepseek_v2.h"            // IWYU pragma: keep
-#include "llm/deepseek_v3.h"            // IWYU pragma: keep
-#include "llm/deepseek_v32.h"           // IWYU pragma: keep
-#include "llm/glm5.h"                   // IWYU pragma: keep
-#include "llm/glm5_mtp.h"               // IWYU pragma: keep
-#include "llm/joyai_llm_flash.h"        // IWYU pragma: keep
-#include "llm/joyai_llm_flash_mtp.h"    // IWYU pragma: keep
-#include "llm/mtp_model_base.h"         // IWYU pragma: keep
-#include "llm/oxygen.h"                 // IWYU pragma: keep
-#include "llm/qwen2.h"                  // IWYU pragma: keep
-#include "llm/qwen3.h"                  // IWYU pragma: keep
-#include "llm/qwen3_moe.h"              // IWYU pragma: keep
-#include "vlm/oxygen_vlm.h"             // IWYU pragma: keep
-#include "vlm/qwen2_5_vl.h"             // IWYU pragma: keep
-#include "vlm/qwen2_vl.h"               // IWYU pragma: keep
-#include "vlm/qwen3_5.h"                // IWYU pragma: keep
-#include "vlm/qwen3_vl.h"               // IWYU pragma: keep
-#include "vlm/qwen3_vl_moe.h"           // IWYU pragma: keep
+#include "dit/pipelines/pipeline_flux.h"          // IWYU pragma: keep
+#include "dit/pipelines/pipeline_flux_control.h"  // IWYU pragma: keep
+#include "dit/pipelines/pipeline_flux_fill.h"     // IWYU pragma: keep
+#include "llm/deepseek_mtp.h"                     // IWYU pragma: keep
+#include "llm/deepseek_v2.h"                      // IWYU pragma: keep
+#include "llm/deepseek_v3.h"                      // IWYU pragma: keep
+#include "llm/deepseek_v32.h"                     // IWYU pragma: keep
+#include "llm/glm5.h"                             // IWYU pragma: keep
+#include "llm/glm5_mtp.h"                         // IWYU pragma: keep
+#include "llm/joyai_llm_flash.h"                  // IWYU pragma: keep
+#include "llm/joyai_llm_flash_mtp.h"              // IWYU pragma: keep
+#include "llm/mtp_model_base.h"                   // IWYU pragma: keep
+#include "llm/oxygen.h"                           // IWYU pragma: keep
+#include "llm/qwen2.h"                            // IWYU pragma: keep
+#include "llm/qwen3.h"                            // IWYU pragma: keep
+#include "llm/qwen3_moe.h"                        // IWYU pragma: keep
+#include "vlm/oxygen_vlm.h"                       // IWYU pragma: keep
+#include "vlm/qwen2_5_vl.h"                       // IWYU pragma: keep
+#include "vlm/qwen2_vl.h"                         // IWYU pragma: keep
+#include "vlm/qwen3_5.h"                          // IWYU pragma: keep
+#include "vlm/qwen3_vl.h"                         // IWYU pragma: keep
+#include "vlm/qwen3_vl_moe.h"                     // IWYU pragma: keep
 #elif defined(USE_ILU)
 #include "llm/qwen2.h"      // IWYU pragma: keep
 #include "llm/qwen3.h"      // IWYU pragma: keep
 #include "llm/qwen3_moe.h"  // IWYU pragma: keep
 #elif defined(USE_CUDA)
-#include "dit/pipeline_longcat_audiodit.h"    // IWYU pragma: keep
-#include "dit/pipeline_longcat_image.h"       // IWYU pragma: keep
-#include "dit/pipeline_longcat_image_edit.h"  // IWYU pragma: keep
-#include "llm/qwen2.h"                        // IWYU pragma: keep
-#include "llm/qwen3.h"                        // IWYU pragma: keep
-#include "llm/qwen3_moe.h"                    // IWYU pragma: keep
-#include "vlm/qwen2_5_vl.h"                   // IWYU pragma: keep
-#include "vlm/qwen2_vl.h"                     // IWYU pragma: keep
-#include "vlm/qwen3_vl.h"                     // IWYU pragma: keep
-#include "vlm/qwen3_vl_moe.h"                 // IWYU pragma: keep
+#include "dit/pipelines/pipeline_longcat_audiodit.h"    // IWYU pragma: keep
+#include "dit/pipelines/pipeline_longcat_image.h"       // IWYU pragma: keep
+#include "dit/pipelines/pipeline_longcat_image_edit.h"  // IWYU pragma: keep
+#include "llm/qwen2.h"                                  // IWYU pragma: keep
+#include "llm/qwen3.h"                                  // IWYU pragma: keep
+#include "llm/qwen3_moe.h"                              // IWYU pragma: keep
+#include "vlm/qwen2_5_vl.h"                             // IWYU pragma: keep
+#include "vlm/qwen2_vl.h"                               // IWYU pragma: keep
+#include "vlm/qwen3_vl.h"                               // IWYU pragma: keep
+#include "vlm/qwen3_vl_moe.h"                           // IWYU pragma: keep
 #elif defined(USE_MUSA)
 #include "llm/musa/qwen3.h"  // IWYU pragma: keep
 #else
